+++ packages/yak_runner/lib/src/delegates/test/factory/unary.dart
@@ -1,8 +1,8 @@
 import '../delegate/all.dart';
 
-/// a `mixin` that allow to generate `YakRunnerTestDelegate`
+/// a `mixin` that allow to generate `RunnerTestDelegate`
 /// of the same types `T` & `S` of the the class calling the function
-mixin YakRunnerArgTestFactoryDelegate {
-  /// `call` returns a `YakRunnerArgTestDelegate` with types `T` &  `S`
-  YakRunnerArgTestDelegate<T, S> call<T, S>();
+mixin UnaryRunnerTestFactoryDelegate {
+  /// `call` returns a `UnaryRunnerTestDelegate` with types `T` &  `S`
+  UnaryRunnerTestDelegate<T, S> call<T, S>();
 }