+++ packages/yak_runner/lib/src/extensions/all.dart
@@ -1,2 +1,4 @@
 export 'custom/all.dart';
+export 'failure/all.dart';
+export 'iterable/all.dart';
 export 'result/all.dart';