+++ packages/yak_runner/lib/src/extensions/failure/all.dart
@@ -1,0 +1,1 @@
+export 'of_type.dart';