+++ packages/yak_runner/lib/src/mixin/test/unary.dart
@@ -1,11 +1,11 @@
 import '../../all.dart';
 
-/// a `mixin` on `YakRunnerBase`
-mixin YakRunnerArgTestMixin<T, S> on YakRunnerBase {
-  /// is a `YakRunnerTestDelegate Function(YakRunnerTestFactoryDelegate)`
+/// a `mixin` on `RunnerBase`
+mixin UnaryRunnerTestMixin<T, S> on RunnerBase {
+  /// is a `RunnerTestDelegate Function(RunnerTestFactoryDelegate)`
   /// types `T` & `S` are passed by the `Object`
-  /// to the test to `YakRunnerTestDelegate`
-  YakRunnerArgTestDelegate<T, S> buildTest(
-          YakRunnerArgTestFactoryDelegate testFactory) =>
+  /// to the test to `RunnerTestDelegate`
+  UnaryRunnerTestDelegate<T, S> buildTest(
+          UnaryRunnerTestFactoryDelegate testFactory) =>
       testFactory<T, S>();
 }