+++ packages/yak_runner/lib/src/typedef/all.dart
@@ -1,2 +1,3 @@
+export 'delegate/all.dart';
 export 'failure_of_type.dart';
 export 'handle/all.dart';