+++ packages/yak_runner/test/mixin/test/test_delegate/nullary.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:yak_runner/yak_runner.dart';
 
-class FakeNullaryTest<T> implements YakRunnerTestDelegate<T> {
+class FakeNullaryTest<T> implements RunnerTestDelegate<T> {
   const FakeNullaryTest();
   @override
   void call(FutureOr<T> result) {}