+++ packages/yak_runner/test/mixin/test/test_delegate/unary.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:yak_runner/yak_runner.dart';
 
-class FakeUnaryTest<T, S> implements YakRunnerArgTestDelegate<T, S> {
+class FakeUnaryTest<T, S> implements UnaryRunnerTestDelegate<T, S> {
   const FakeUnaryTest();
 
   @override