+++ packages/yak_runner/test/mocks/delegates/test/factory/nullary.dart
@@ -1,8 +1,8 @@
 import 'package:stub/stub.dart';
 import 'package:yak_runner/yak_runner.dart';
 
-class MockYakRunnerTestFactoryDelegate extends Stub<dynamic>
-    implements YakRunnerTestFactoryDelegate {
+class MockRunnerTestFactoryDelegate extends Stub<dynamic>
+    implements RunnerTestFactoryDelegate {
   @override
-  YakRunnerTestDelegate<T> call<T>() => stub();
+  RunnerTestDelegate<T> call<T>() => stub();
 }