+++ packages/yak_test/test/async/unary_test.dart
@@ -1,7 +1,7 @@
 import 'package:yak_runner/yak_runner.dart';
 import 'package:yak_test/yak_test.dart';
 
-final runner = YakRunnerArgAsync<String, int>((i) async => '$i');
+final runner = UnaryRunnerAsync<String, int>((i) async => '$i');
 final _test = runner.buildTest(
   runner.buildTestFactory('unary test'),
 );